% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{age_components}
\alias{age_components}
\title{Extract age components from a model}
\usage{
age_components(object, ...)
}
\arguments{
\item{object}{A vital mable object with a single model column.}

\item{...}{Not currently used.}
}
\value{
vital object containing the age components from the model.
}
\description{
For a mable with a single model column, return the model components
that are indexed by age.
}
\examples{
aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(lee_carter = LC(log(Mortality))) |>
  age_components()

}
