% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_tsibble.R
\name{as_life_tsibble}
\alias{as_life_tsibble}
\alias{as_life_tsibble.tbl_ts}
\title{Coerce to a life_tsibble object}
\usage{
as_life_tsibble(x, ...)

\method{as_life_tsibble}{tbl_ts}(
  x,
  age,
  sex = NULL,
  deaths = NULL,
  births = NULL,
  population = NULL
)
}
\arguments{
\item{x}{Objects to be coerced to a life_tsibble (`life_tbl_ts`).}

\item{...}{Other arguments passed on to individual methods.}

\item{age}{Character string specifying name of age variable (required)}

\item{sex}{Character string specifying name of sex variable (optional)}

\item{deaths}{Character string specifying name of deaths variable (optional)}

\item{births}{Character string specifying name of births variable (optional)}

\item{population}{Character string specifying name of population variable (optional)}

\item{sex_groups}{Logical variable indicating of the groups denote sexes}

\item{validate}{`TRUE` suggests to verify that each key or each combination
of key variables leads to unique time indices (i.e. a valid tsibble). If you
are sure that it's a valid input, specify `FALSE` to skip the checks.}
}
\value{
A life_tsibble object.
}
\description{
Coerce to a life_tsibble object
}
\examples{
# coerce demogdata object to life_tsibble
as_life_tsibble(demography::fr.mort)
}
\seealso{
\code{\link[tsibble]{tsibble}()}
}
