% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HU.R
\name{autoplot.fpc_model}
\alias{autoplot.fpc_model}
\title{Plot of functional data model components}
\usage{
\method{autoplot}{fpc_model}(object, show_order = 2, ...)
}
\arguments{
\item{object}{output from \code{\link{fpc_model}}.}

\item{show_order}{The maximum number of basis functions to plot}

\item{...}{Other arguments not used.}
}
\value{
A patchwork object
}
\description{
Plot of mean, basis functions and PC scores from a functional data model
}
