% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_carter.R
\name{autoplot.lca_model}
\alias{autoplot.lca_model}
\title{Plot of Lee-Carter model components}
\usage{
\method{autoplot}{lca_model}(object, ...)
}
\arguments{
\item{object}{output from \code{\link{lee_carter}}.}

\item{...}{Other arguments not used.}
}
\value{
A list of ggplot objects.
}
\description{
Plot of ax, bx and kt from a Lee-Carter model.
}
