% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{AVERAGE}
\alias{AVERAGE}
\alias{report.model_ave}
\title{Mean models}
\usage{
AVERAGE(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{AVERAGE()} returns an iid model applied to the formula's response variable.
}
\examples{
library(dplyr)
aus_mortality |>
  filter(State == "Victoria", Sex == "female") |>
  model(mean = AVERAGE(Mortality))
}
\seealso{
[Forecasting: Principles and Practices, Some simple forecasting methods (section 3.2)](https://otexts.com/fpp3/simple-methods.html)
}
