% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmean.R
\name{fitted.model_fmean}
\alias{fitted.model_fmean}
\title{Fitted values from a mean model}
\usage{
\method{fitted}{model_fmean}(object, ...)
}
\arguments{
\item{object}{TBD}

\item{...}{Not used.}
}
\description{
Fitted values from a mean model
}
