% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast.lca_model}
\alias{forecast.lca_model}
\title{Forecasts using Lee-Carter method.}
\usage{
\method{forecast}{lca_model}(
  object,
  h = 50,
  se = c("innovdrift", "innovonly"),
  jumpchoice = c("fit", "actual"),
  level = 80,
  ...
)
}
\arguments{
\item{object}{Output from \code{\link{lee_carter}}.}

\item{h}{Number of years ahead to forecast.}

\item{se}{Method used for computation of standard error.
Possibilities: \dQuote{innovdrift} (innovations and drift) and \dQuote{innovonly} (innovations only).}

\item{jumpchoice}{Method used for computation of jumpchoice.
Possibilities: \dQuote{actual} (use actual rates from final year) and \dQuote{fit} (use fitted rates).
The original Lee-Carter method used 'fit' (the default), but Lee and Miller (2001)
and most other authors prefer 'actual' (the default).}

\item{level}{Confidence level for prediction intervals.}

\item{...}{Other arguments.}
}
\value{
Object of class \code{fm_forecast} with the following components:
\item{age}{Ages from \code{object}.}
\item{year}{Years from \code{object}.}
\item{rate}{List of matrices containing forecasts, lower bound and upper bound of prediction intervals.
  Point forecast matrix takes the same name as the series that has been forecast.}
\item{fitted}{Matrix of one-step forecasts for historical data}
Other components included are
\item{e0}{Forecasts of life expectancies (including lower and upper bounds)}
\item{kt.f}{Forecasts of coefficients from the model.}
\item{type}{Data type.}
\item{model}{Details about the fitted model}
}
\description{
The kt coefficients are forecast using a random walk with drift.
The forecast coefficients are then multiplied by bx to obtain a forecast
demographic rate curve.
}
\examples{
library(dplyr)
library(ggplot2)
ausf_lca <- aus_mortality |>
  filter(Sex == "female", State == "Australia") |>
  lee_carter()
ausf_fcast <- forecast(ausf_lca, 50)
ausf_fcast |>
  filter(Age == 60) |>
  autoplot(aus_mortality)
rainbow_plot(ausf_fcast, .mean) +
  scale_y_log10()

}
\references{
Lee, R D, and Carter, L R (1992) Modeling and forecasting US mortality.
\emph{Journal of the American Statistical Association}, \bold{87}, 659-671.

Lee R D, and Miller T (2001). Evaluating the performance of the Lee-Carter
  method for forecasting mortality. \emph{Demography}, \bold{38}(4), 537–549.
}
\author{
Rob J Hyndman
}
\keyword{models}
