% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HU.R
\name{fpc_model}
\alias{fpc_model}
\title{Fit a functional principal components model}
\usage{
fpc_model(.data, rates, order = 6)
}
\arguments{
\item{.data}{A vital object including an age variable and a variable containing mortality or fertility rates.}

\item{rates}{Variable in `.data` containing mortality or fertility rates.
If omitted, it will search for a variable with one of the following names:
`mx`, `mortality`, `fx`, `fertility` or `rate` (not case sensitive).}

\item{order}{Number of principal components to fit.}
}
\value{
A list of class \code{fpc_model} containing various model objects.
}
\description{
Model mortality or fertility data using functional principal components
as proposed by Hyndman & Ullah (2007)
Missing rates are set to the geometric mean rate for the relevant age.
}
\examples{
# Compute Hyndman-Ullah model for Australian females, males and total
aus_hu <- aus_mortality |>
  dplyr::filter(Code == "AUS") |>
  fpc_model()
aus_hu
autoplot(aus_hu) +
  patchwork::plot_annotation("Hyndman-Ullah components for Australia")
}
\references{
Hyndman, R.J., and Ullah, S. (2007) Robust forecasting of
mortality and fertility rates: a functional data approach.
\emph{Computational Statistics & Data Analysis}, \bold{51}, 4942-4956.
\url{https://robjhyndman.com/publications/funcfor/}
}
\author{
Rob J Hyndman
}
\keyword{models}
