% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{generate.model_ave}
\alias{generate.model_ave}
\title{Generate from a mean model}
\usage{
\method{generate}{model_ave}(x, new_data, bootstrap = FALSE, ...)
}
\arguments{
\item{x}{A mable.}

\item{new_data}{A `tsibble` containing future information used to forecast.}

\item{bootstrap}{If `TRUE`, then forecast distributions are computed using simulation with resampled errors.}

\item{...}{Additional arguments not used.}
}
\description{
Generate from a mean model
}
