% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R
\name{glance.model_lc}
\alias{glance.model_lc}
\title{Glance a average method model}
\usage{
\method{glance}{model_lc}(x, ...)
}
\arguments{
\item{x}{model or other R object to convert to single-row data frame}

\item{...}{other arguments passed to methods}
}
\value{
A one row tibble summarising the model's fit.
}
\description{
Construct a single row summary of the average method model.
}
\details{
Contains the variance of residuals (`sigma2`).
}
