% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_hmd.R
\name{hmd_data}
\alias{hmd_data}
\alias{hmd_life_expectancy}
\alias{hmd_birth}
\alias{hmd_life_table}
\alias{hmd_population}
\title{Fetch data from the Human Mortality Database}
\usage{
hmd_data(country, stats = "death_rate", year_range = 1, age_range = 1)

hmd_life_expectancy(country, year_range = 1)

hmd_birth(country)

hmd_life_table(country, age_range = 1, year_range = 1)

hmd_population(country, age_range = 1)
}
\arguments{
\item{country}{The country code to fetch the data for. The input may be
a named or unnamed character vector. If the vector is named, then the
name is used as the label.}

\item{stats}{The statistics to extract. The available statistics are:
"birth", "death", "life_expectancy", "exposure_to_risk", "population",
"life_tables", and "death_rate".}

\item{year_range, age_range}{A single integer indicating the period or
cohort year/age range. Generally, only 1, 5 or 10 are available.}
}
\value{
A humble object.
}
\description{
The `hmd_data()` fetches data from multiple statistics and merge them together
in the output. All other functions grab a particular (set of) statistics from
the Human Mortality Database (mortality.org). You need to register an account at
mortality.org to get the data and call the `hmd_session()` first before
using these functions.
}
\seealso{
hmd_session
}
