% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_hmd.R
\name{hmd_session}
\alias{hmd_session}
\title{Set the username and password to access the former site of the
Human Mortality Database}
\usage{
hmd_session(
  username = Sys.getenv("HMD_USERNAME"),
  password = Sys.getenv("HMD_PASSWORD")
)
}
\arguments{
\item{username}{The username (typically your email) that you signed up for HMD.}

\item{password}{The password used to access HMD.}
}
\description{
To download data from the HMD using this package, you need to register at
the "former" HMD website (https://former.mortality.org),
Usernames and passwords for the new HMD site (https://mortality.org) will not
work. This is a temporary patch until an API is released for the new site.
}
\details{
You can either pass the username and password as arguments to the function,
or you can store your username and password in the .Renviron file
containing

HMD_USERNAME=<YOUR USERNAME>

HMD_PASSWORD=<YOUR PASSWORD>
}
\author{
Emi Tanaka
}
