% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_expectancy.R
\name{life_expectancy}
\alias{life_expectancy}
\title{Compute life expectancy from mortality rates}
\usage{
life_expectancy(.data, from_age = 0, age, sex, mortality)
}
\arguments{
\item{.data}{A tsibble including an age variable and a variable containing mortality rates.}

\item{from_age}{Age at which life expectancy to be calculated. Either a scalar or a vector of ages.}

\item{age}{Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` or `Age_group` will be used (not case sensitive).}

\item{sex}{Optional variable in `.data` containing sex information. If omitted, the variable with name `Sex` or `Group` will be used (not case sensitive).}

\item{mortality}{Variable in `.data` containing Mortality rates (mx). If omitted, the variable with name  `mx`, `Mortality` or `Rate` will be used (not case sensitive).}
}
\value{
A tsibble object with life expectancy in column `ex`.
}
\description{
Returns remaining life expectancy at a given age (0 by default).
}
\examples{
# Compute Australia life expectancy for females over time
library(dplyr)
aus_mortality |>
  filter(Code=="AUS", Sex=="female") |>
  life_expectancy()
}
\references{
Chiang CL. (1984) \emph{The life table and its applications}. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) \emph{Applied mathematical demography}, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) \emph{Demography: measuring and modeling population processes}. Blackwell
}
\seealso{
life_table
}
