% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway.R
\name{norway_births}
\alias{norway_births}
\title{Norwegian mortality and births data}
\format{
Time series of class \code{vital}
}
\source{
Human Mortality Database \url{https://mortality.org}
}
\description{
\code{norway_births} is an annual \code{vital} object covering the years 1846-2022, as provided
by the Human Mortality Database on 21 April 2024.
}
\examples{
library(ggplot2)
norway_births
norway_births |>
  autoplot(Births)

}
\keyword{datasets}
