% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norway.R
\name{norway_mortality}
\alias{norway_mortality}
\title{Norwegian mortality data}
\format{
Time series of class \code{vital}
}
\source{
Human Mortality Database \url{https://mortality.org}
}
\description{
\code{norway_mortality} is an annual \code{vital} covering the years 1846-2022, as provided
by the Human Mortality Database on 21 April 2024.
}
\examples{
library(ggplot2)
norway_mortality
norway_mortality |>
  dplyr::filter(Age < 85, Year < 1900, Sex != "Total") |>
  autoplot(Mortality) +
  scale_y_log10()

}
\keyword{datasets}
