% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hfd.R
\name{read_hfd_files}
\alias{read_hfd_files}
\title{Read data from files downloaded from HFD and construct a \code{vital} object for use in other functions}
\usage{
read_hfd_files(files)
}
\arguments{
\item{files}{Vector of file names containing data downloaded from the HFD.
The file names are used to determine what they contain. If the file names are
as per the HFD, then the function will automatically determine the contents.
If it is unclear what a file contains, the columns will be named according to the filename.
If the data contains a mixture of age-specific and non-age-specific variables,
then the non-age-specific data will be repeated for each age. If you have HFD files
for many countries, all with the same names, then you should put them in separate
folders to avoid confusion, and to save changing all the filenames.}
}
\value{
\code{read_hfd_files} returns a \code{vital} object combining the downloaded data.
}
\description{
\code{read_hfd_files} reads single-year and single-age data from files downloaded from the Human Fertility
Database (HFD \url{https://www.humanfertility.org}) and constructs a \code{vital} object suitable
for use in other functions. This function uses \code{\link[HMDHFDplus:readHFD]{HMDHFDplus::readHFD()}}
to parse the files.
}
\examples{
\dontrun{
# File downloaded from the [Human Fertility Database](https://www.humanfertility.org)
fertility <- read_hfd_files("NORasfrRR.txt")
}
}
\author{
Rob J Hyndman
}
\keyword{manip}
