% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{read_hmd}
\alias{read_hmd}
\title{Read text file from HMD}
\usage{
read_hmd(file)
}
\arguments{
\item{file}{A path to a file, or a connection.}
}
\value{
A tsibble with an annual index and (possibly) Age and Sex keys.
}
\description{
Read text files from the Human Mortality Database (\url{http://mortality.org}).
It currently handles births, deaths, population, exposure, death rates,
life tables and life expectancy files.
It will not work with cohort data or lexis triangles.
}
\examples{
\dontrun{
aus <- read_hmd("AUS_Mx_1x1.txt")
}
}
\seealso{
\code{\link[HMDHFDplus]{readHMD}}.
}
