% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_births.R
\name{read_hmd_births}
\alias{read_hmd_births}
\title{Read data directly from HMD and construct a \code{vital} object for use in other functions}
\usage{
read_hmd_births(country, username, password)
}
\arguments{
\item{country}{Directory abbreviation from the HMD. For instance, Australia = "AUS".}

\item{username}{HMD username (case-sensitive)}

\item{password}{HMD password (case-sensitive)}
}
\value{
\code{read_hmd_births} returns a \code{vital} object with the following variables:
\item{Year}{Numerical vector containing year of observation}
\item{Sex}{Character vector taking values "Female", "Male", "Total"}
\item{Births}{Total live births}
}
\description{
\code{read_hmd_births} reads single year Births data from the Human Mortality Database (HMD
\url{https://www.mortality.org}) and constructs a \code{vital} object suitable
for use in other functions. This function uses \code{\link[HMDHFDplus:readHMDweb]{HMDHFDplus::readHMDweb()}}
to download the required data. To read handle Deaths, Population, Exposure,
and Death Rates, use either \code{read_hmd()} or \code{read_hmd_files()}.
}
\details{
In order to read the data, users are required to create an account with the
HMD website (\url{https://www.mortality.org}), and obtain a valid username and password.
}
\examples{
\dontrun{
norway <- read_hmd_births(
  country = "NOR",
  username = "Nora.Weigh@mymail.com",
  password = "FF!5xeEFa6"
)
}
}
\author{
Rob J Hyndman
}
