% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd_births.R
\name{read_hmd_births_file}
\alias{read_hmd_births_file}
\title{Read data from files downloaded from HMD and construct a \code{vital} object for use in other functions}
\usage{
read_hmd_births_file(filename)
}
\arguments{
\item{filename}{Name of the file containing the data.}
}
\value{
\code{read_hmd_births_file} returns a \code{vital} object containing the following variables:
\item{Year}{Numerical vector containing year of observation}
\item{Sex}{Character vector taking values "Female", "Male", "Total"}
\item{Births}{Total live births}
}
\description{
\code{read_hmd_births_file} reads single year births data from the appropriate file
downloaded from the Human Mortality Database (HMD \url{https://www.mortality.org})
and constructs a \code{vital} object suitable for use in other functions. This
function uses \code{\link[HMDHFDplus:readHMD]{HMDHFDplus::readHMD()}} to parse the files.
}
\examples{
\dontrun{
read_hmd_births_file("Births.txt")
}
}
\author{
Rob J Hyndman
}
\keyword{manip}
