% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{read_hmd_files}
\alias{read_hmd_files}
\title{Read data from files downloaded from HMD and construct a \code{vital} object for use in other functions}
\usage{
read_hmd_files(Deaths = NULL, Exposures = NULL, Population = NULL, Mx = NULL)
}
\arguments{
\item{Deaths}{File containing deaths data}

\item{Exposures}{File containing exposures data}

\item{Population}{File containing population data}

\item{Mx}{File containing mortality data}
}
\value{
\code{read_hmd_files} returns a \code{vital} object with the following variables:
\item{Year}{Numerical vector containing year of observation}
\item{Age}{Numerical vector containing age group}
\item{OpenInterval}{Logical vector indicating if the age interval is open}
\item{Sex}{Character vector taking values "Female", "Male", "Total"}
\item{Deaths}{Total deaths}
\item{Exposure}{Numerical vector of exposure-to-risk, equal to population size on 30 June}
\item{Population}{Numerical vector containing population size on 1 January}
\item{Mortality}{Numerical vector of mortality rate, equal to Deaths/Exposure}
}
\description{
\code{read_hmd_files} reads 1x1 data from files downloaded from the Human Mortality
Database (HMD \url{https://www.mortality.org}) and constructs a \code{vital} object suitable
for use in other functions. This function uses \code{\link[HMDHFDplus]{readHMD}}
to parse the files. At least one file is required. If any file is missing, the corresponding
variable will be missing from the output.
}
\examples{
\dontrun{
read_hmd_files("Deaths_1x1.txt", "Exposures_1x1.txt", "Population.txt", "Mx_1x1.txt")
}
}
\author{
Rob J Hyndman
}
\keyword{manip}
