% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktdb.R
\name{read_ktdb_file}
\alias{read_ktdb_file}
\title{#' Read old-age mortality data from files downloaded from K-T database}
\usage{
read_ktdb_file(file1, file2 = NULL, triangle = 1, male_first = TRUE)
}
\arguments{
\item{file1}{Name of the first file containing data downloaded from the K-T database.}

\item{file2}{Name of the second file containing data downloaded from the K-T database.}

\item{triangle}{Lexis triangle number, 1 (default) is lower triangle, 2 is upper triangle.}

\item{male_first}{Indicator of whether file1 is for males. Default is TRUE.}
}
\value{
\code{read_ktdb_file} returns a \code{vital} object combining the downloaded data.
}
\description{
\code{read_ktdb_file} reads old-age mortality data from a file downloaded from
K-T database (\url{https://www.demogr.mpg.de/cgi-bin/databases/ktdb/datamap.plx})
and constructs a \code{vital} object suitable for use in other functions.
If two files are provided, the function will treat them as data for each gender,
returning a combined dataset. If only one file is provided, the function will
assume that it represents data for a single gender.
}
\examples{
\dontrun{
# File downloaded from the K-T database
australia_male <- read_ktdb_file("maustl.txt")
}

}
\author{
Sixian Tang
}
