% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMF.R
\name{read_stmf_data}
\alias{read_stmf_data}
\title{Read Short-Term Mortality Fluctuations data from HMD and construct a \code{vital} object for use in other functions}
\usage{
read_stmf_data(country)
}
\arguments{
\item{country}{Directory abbreviation from the HMD. For instance, Australia = "AUS".}
}
\value{
\code{read_stmf_data} returns a \code{vital} object combining the downloaded data.
}
\description{
\code{read_stmf} reads weekly mortality data from the Short-term Mortality Fluctuations (STMF)
series available in the Human Mortality Database (HMD) \url{https://www.mortality.org/Data/STMF})
and constructs a \code{vital} object suitable for use in other functions.
}
\examples{
\dontrun{
norway <- read_stmf_data(country = "NOR")
}


}
\author{
Sixian Tang
}
