% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{forecast}
\alias{generate}
\alias{model}
\alias{report}
\alias{interpolate}
\alias{tidy}
\alias{augment}
\alias{estimate}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabletools}{\code{\link[fabletools:augment]{augment}}, \code{\link[fabletools:estimate]{estimate()}}, \code{\link[fabletools:forecast]{forecast}}, \code{\link[fabletools:generate]{generate}}, \code{\link[fabletools:glance]{glance}}, \code{\link[fabletools:interpolate]{interpolate}}, \code{\link[fabletools:model]{model()}}, \code{\link[fabletools:report]{report()}}, \code{\link[fabletools:tidy]{tidy}}}

  \item{ggplot2}{\code{\link[ggplot2:autoplot]{autoplot()}}}
}}

