% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{refit.model_ave}
\alias{refit.model_ave}
\title{Refit a AVERAGE model}
\usage{
\method{refit}{model_ave}(object, new_data, reestimate = FALSE, ...)
}
\arguments{
\item{object}{A mable}

\item{new_data}{A dataset with the same structure as the data used to fit the model.}

\item{reestimate}{If `TRUE`, the mean for the fitted model will be re-estimated
to suit the new data.}

\item{...}{Additional arguments not used.}
}
\description{
Applies a fitted average method model to a new dataset.
}
\examples{
lung_deaths_male <- as_tsibble(mdeaths)
lung_deaths_female <- as_tsibble(fdeaths)

fit <- lung_deaths_male \%>\%
  model(AVERAGE(value))

report(fit)

fit \%>\%
  refit(lung_deaths_female) \%>\%
  report()
}
