% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R
\name{residuals.model_lc}
\alias{residuals.model_lc}
\title{Residuals from a mean model}
\usage{
\method{residuals}{model_lc}(object, ...)
}
\arguments{
\item{object}{TBD}

\item{...}{Not used.}
}
\description{
Residuals from a mean model
}
