% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_spline}
\alias{smooth_spline}
\alias{smooth_mortality}
\alias{smooth_fertility}
\alias{smooth_loess}
\title{Functions to smooth demographic data}
\usage{
smooth_spline(.data, .var, k = -1, weights = NULL)

smooth_mortality(.data, .var, b = 65, power = 0.4, k = 30, weights = NULL)

smooth_fertility(.data, .var, lambda = 1e-10)

smooth_loess(.data, .var, span = 0.2, weights = NULL)
}
\arguments{
\item{.data}{A vital object}

\item{.var}{name of variable to smooth}

\item{k}{Number of knots to use for penalized regression spline estimate.}

\item{weights}{Vector of weights.}

\item{b}{Lower age for monotonicity. Above this, the smooth curve is assumed to be monotonically increasing.}

\item{power}{Power transformation for age variable before smoothing. Default is 0.4 for mortality data and 1 (no transformation) for fertility or migration data.}

\item{lambda}{Penalty for constrained regression spline.}

\item{span}{Span for loess smooth.}
}
\value{
vital with added columns containing smoothed values and their standard errors
}
\description{
These smoothing functions allow smoothing of a variable in a vital object.
The vital object is returned along with some additional columns containing
information about the smoothed variable: usually `.smooth` containing the
smoothed values, and `.smooth_se` containing the corresponding standard errors.
}
\details{
`smooth_mortality()` use penalized regression splines applied to log mortality
with a monotonicity constraint above age `b`. The methodology is based on Wood (1994).
`smooth_fertility()` uses weighted regression B-splines with a concavity constraint,
based on He and Ng (1999). The function `smooth_loess()` uses locally quadratic
regression, while `smooth_spline()` uses penalized regression splines.
}
\examples{
library(dplyr)
aus_mortality |>
	filter(State == "Victoria", Sex == "female") |>
	smooth_mortality(Mortality)
aus_fertility |>
	smooth_fertility(Fertility)
}
\author{
Rob J Hyndman
}
\keyword{smooth}
