% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R, R/fmean.R
\name{tidy.LC}
\alias{tidy.LC}
\alias{tidy.FMEAN}
\title{Extract model coefficients from a mable}
\usage{
\method{tidy}{LC}(x, ...)

\method{tidy}{FMEAN}(x, ...)
}
\arguments{
\item{x}{A mable.}

\item{...}{Arguments for model methods.}
}
\description{
This function will obtain the coefficients (and associated statistics) for
each model in the mable.

This function will obtain the coefficients (and associated statistics) for
each model in the mable.
}
\author{
Rob J Hyndman
}
