% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{tidylife_conflicts}
\alias{tidylife_conflicts}
\title{Conflicts between tidylife packages and other packages}
\usage{
tidylife_conflicts()
}
\value{
A list object of class \code{tidylife_conflicts}.
}
\description{
This function lists all the conflicts between packages in the tidylife collection
and other packages that you have loaded.
}
\details{
Some conflicts are deliberately ignored: \code{intersect}, \code{union},
\code{setequal}, and \code{setdiff} from dplyr; and \code{intersect},
\code{union}, \code{setdiff}, and \code{as.difftime} from lubridate.
These functions make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
tidylife_conflicts()
}
