% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tidylife_packages}
\alias{tidylife_packages}
\title{List all packages loaded by tidylife}
\usage{
tidylife_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include tidylife in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages loaded by tidylife
}
\examples{
tidylife_packages()
}
