% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{vital}
\alias{vital}
\title{Create a vital object}
\usage{
vital(
  ...,
  key = NULL,
  index,
  .age = NULL,
  .sex = NULL,
  .deaths = NULL,
  .births = NULL,
  .population = NULL,
  regular = TRUE,
  .drop = TRUE
)
}
\arguments{
\item{...}{A set of name-value pairs}

\item{key}{Variable(s) that uniquely determine time indices. NULL for empty key,
and \code{c()} for multiple variables. It works with tidy selector
(e.g. \code{\link{dplyr::starts_with}()}).}

\item{index}{A variable to specify the time index variable.}

\item{.age}{Character string specifying name of age variable}

\item{.sex}{Character string specifying name of sex variable}

\item{.deaths}{Character string specifying name of deaths variable}

\item{.births}{Character string specifying name of births variable}

\item{.population}{Character string specifying name of population variable}

\item{regular}{Regular time interval (TRUE) or irregular (FALSE). The interval
is determined by the greatest common divisor of index column, if TRUE.}

\item{.drop}{If TRUE, empty key groups are dropped.}
}
\description{
A vital object is a type of tsibble that contains vital statistics such as
births, deaths, and population counts, and mortality and fertility rates.
It is a tsibble with a special class that allows for special methods to be used.
The object has an attribute that stores variables names needed for some functions,
including age, sex, births, deaths and population.
}
\seealso{
\code{\link[tsibble]{tsibble}()}
}
