% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vital_packages}
\alias{vital_packages}
\title{List all packages loaded by vital}
\usage{
vital_packages(include_self = FALSE)
}
\arguments{
\item{include_self}{Include vital in the list?}
}
\value{
A character vector of package names.
}
\description{
List all packages loaded by vital
}
\examples{
vital_packages()
}
