% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{as_vital}
\alias{as_vital}
\alias{as_vital.demogdata}
\alias{as_vital.tbl_ts}
\title{Coerce to a vital object}
\usage{
as_vital(x, ...)

\method{as_vital}{demogdata}(x, sex_groups = TRUE, ..., validate = TRUE)

\method{as_vital}{tbl_ts}(
  x,
  age,
  sex = NULL,
  deaths = NULL,
  births = NULL,
  population = NULL,
  ...
)
}
\arguments{
\item{x}{Objects to be coerced to a vital format.}

\item{...}{Other arguments passed on to individual methods.}

\item{sex_groups}{Logical variable indicating of the groups denote sexes}

\item{validate}{`TRUE` suggests to verify that each key or each combination
of key variables leads to unique time indices (i.e. a valid tsibble). If you
are sure that it's a valid input, specify `FALSE` to skip the checks.}

\item{age}{Character string specifying name of age variable (required)}

\item{sex}{Character string specifying name of sex variable (optional)}

\item{deaths}{Character string specifying name of deaths variable (optional)}

\item{births}{Character string specifying name of births variable (optional)}

\item{population}{Character string specifying name of population variable (optional)}
}
\value{
A vital object.
}
\description{
A vital object is a type of tsibble that contains vital statistics such as births, deaths, and population counts, and mortality and fertility rates. It is a tsibble with a special class attribute that allows for special methods to be used.
}
\examples{
# coerce demogdata object to vital
as_vital(demography::fr.mort)
}
\seealso{
\code{\link[tsibble]{tsibble}()}
}
