% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital.R
\name{as_vital}
\alias{as_vital}
\alias{as_vital.demogdata}
\alias{as_vital.tbl_ts}
\alias{as_vital.data.frame}
\title{Coerce to a vital object}
\usage{
as_vital(x, ...)

\method{as_vital}{demogdata}(x, sex_groups = TRUE, ...)

\method{as_vital}{tbl_ts}(
  x,
  .age = NULL,
  .sex = NULL,
  .deaths = NULL,
  .births = NULL,
  .population = NULL,
  ...
)

\method{as_vital}{data.frame}(
  x,
  key = NULL,
  index,
  .age = NULL,
  .sex = NULL,
  .deaths = NULL,
  .births = NULL,
  .population = NULL,
  ...
)
}
\arguments{
\item{x}{Object to be coerced to a vital format.}

\item{...}{Other arguments passed to \code{\link[tsibble]{as_tsibble}}}

\item{sex_groups}{Logical variable indicating if the groups denote sexes}

\item{.age}{Character string specifying name of age variable}

\item{.sex}{Character string specifying name of sex variable}

\item{.deaths}{Character string specifying name of deaths variable}

\item{.births}{Character string specifying name of births variable}

\item{.population}{Character string specifying name of population variable}

\item{key}{Variable(s) that uniquely determine time indices. NULL for empty key, and \code{c()} for multiple variables. It works with tidy selector (e.g. \code{\link{dplyr::starts_with}()}).}

\item{index}{A variable to specify the time index variable.}
}
\value{
A vital object.
}
\description{
A vital object is a type of tsibble that contains vital statistics such as births, deaths, and population counts, and mortality and fertility rates.
It is a tsibble with a special class that allows for special methods to be used.
The object has an attribute that stores variables names needed for some functions, including age, sex, births, deaths and population.
}
\examples{
# coerce demogdata object to vital
as_vital(demography::fr.mort)
}
\seealso{
\code{\link[tsibble]{tsibble}()}
}
