% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/australia.R
\name{aus_mortality}
\alias{aus_mortality}
\title{Australian mortality data}
\format{Time series of class `tsibble`}
\source{
Australian Human Mortality Database
}
\description{
\code{aus_mortality} is an annual `tsibble` with three values:
\tabular{ll}{
    Mortality: \tab Mortality rate \cr
    Exposure:  \tab Population at 30 June each year\cr
    Deaths:    \tab Number of deaths \cr
}
}
\details{
The data is disaggregated using three keys:
\tabular{ll}{
    Age:   \tab Age at death\cr
    Sex:   \tab male or female \cr
    State: \tab State of Australia\cr
}
}
\examples{

aus_mortality

}
\keyword{datasets}
