% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.vital}
\alias{autoplot.vital}
\title{Rainbow plot of demographic data against age}
\usage{
\method{autoplot}{vital}(object, ...)
}
\arguments{
\item{.data}{A vital, tsibble or fable object including an age variable and the variable you wish to plot.}

\item{.vars}{A bare expression containing the name of the variable you wish to plot.}

\item{age}{Variable in `.data` containing start year of age intervals.
If omitted, and `.data` is not a vital object, the variable with name `Age` or `Age_group`
will be used (not case sensitive).}
}
\value{
A ggplot2 object.
}
\description{
Produce rainbow plot (coloured by time index) of demographic variable against
against age.
}
\examples{
autoplot(aus_fertility, Fertility)
}
\references{
Hyndman, Rob J & Shang, Han Lin (2010) Rainbow plots, bagplots,
and boxplots for functional data. \emph{Journal of Computational and Graphical Statistics},
\bold{19}(1), 29-45.
}
\author{
Rob J Hyndman
}
