% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_ages.R
\name{collapse_ages}
\alias{collapse_ages}
\title{Collapse upper ages into a single age group. Counts are summed while
rates are recomputed.}
\usage{
collapse_ages(.data, max_age = 100)
}
\arguments{
\item{.data}{A tsibble including an age variable and a variable containing mortality rates.}

\item{max_age}{Maximum age to include in the collapsed age group.}
}
\value{
A tsibble with the same variables as `.data` but with the upper ages collapsed into a single age group.
}
\description{
Collapse upper ages into a single age group. Counts are summed while
rates are recomputed.
}
\examples{
aus_mortality |>
 dplyr::filter(State == "Victoria", Sex == "female") |>
 collapse_ages(max_age = 85)
}
