% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDM.R
\name{FDM}
\alias{FDM}
\alias{report.FDM}
\title{Model mortality or fertility data using functional data approach}
\usage{
FDM(formula, order = 6, ts_model_fn = fable::ARIMA, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{order}{Number of principal components to fit.}

\item{ts_model_fn}{Univariate time series modelling function for the coefficients.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Functional data model of mortality or fertility rates.
\code{FDM()} returns a functional data model applied to the formula's response
variable as a function of age.
}
\examples{
aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(hyndman_ullah = FDM(Mortality)) |>
  report()
}
\references{
Hyndman, R.J., and Ullah, S. (2007) Robust forecasting of
mortality and fertility rates: a functional data approach.
\emph{Computational Statistics & Data Analysis}, \bold{51}, 4942-4956.
\url{https://robjhyndman.com/publications/funcfor/}
}
\author{
Rob J Hyndman
}
