% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmean.R
\name{FMEAN}
\alias{FMEAN}
\alias{report.model_fmean}
\title{Mean models}
\usage{
FMEAN(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{FMEAN()} returns an iid functional model applied to the formula's response variable.
}
\examples{
library(dplyr)
aus_mortality |>
  filter(State == "Victoria", Sex == "female") |>
  model(mean = FMEAN(Mortality))
}
