% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmean.R
\name{FMEAN}
\alias{FMEAN}
\alias{report.FMEAN}
\title{Functional mean models}
\usage{
FMEAN(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{FMEAN()} returns an iid functional model applied to the formula's response variable as a function of age.
}
\examples{
aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(mean = FMEAN(Mortality)) |>
  report()
}
\author{
Rob J Hyndman
}
