% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmean.R
\name{FMEAN}
\alias{FMEAN}
\alias{report.FMEAN}
\title{Functional mean model}
\usage{
FMEAN(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{FMEAN()} returns an iid functional model applied to the formula's response variable as a function of age.
}
\examples{
fmean <- norway_mortality |>
  dplyr::filter(Sex == "Female") |>
  model(mean = FMEAN(Mortality))
report(fmean)
autoplot(fmean) + ggplot2::scale_y_log10()
}
\author{
Rob J Hyndman
}
