% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnaive.R
\name{FNAIVE}
\alias{FNAIVE}
\alias{report.FNAIVE}
\title{Functional naive model}
\usage{
FNAIVE(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used. An error is given if any arguments are supplied here,
so that misspelled arguments are not silently ignored.}
}
\value{
A model specification.
}
\description{
\code{FNAIVE()} returns a random walk functional model applied to the formula's response variable as a function of age.
Standard deviations that cannot be estimated, such as at ages with fewer
than two finite residuals, are interpolated from neighbouring ages.
}
\examples{
fnaive <- norway_mortality |>
  dplyr::filter(Sex == "Female") |>
  model(fit = FNAIVE(Mortality))
report(fnaive)
autoplot(fnaive) + ggplot2::scale_y_log10()
}
\author{
Rob J Hyndman
}
