% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnaive.R
\name{FNAIVE}
\alias{FNAIVE}
\alias{report.FNAIVE}
\title{Functional mean models}
\usage{
FNAIVE(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{FNAIVE()} returns an random walk functional model applied to the formula's response variable as a function of age.
}
\examples{
aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(mean = FNAIVE(Mortality)) |>
  report()
}
\author{
Rob J Hyndman
}
