% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnaive.R
\name{FNAIVE}
\alias{FNAIVE}
\alias{report.FNAIVE}
\title{Functional naive model}
\usage{
FNAIVE(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{FNAIVE()} returns an random walk functional model applied to the formula's response variable as a function of age.
}
\examples{
fnaive <- aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(fit = FNAIVE(Mortality))
report(fnaive)
autoplot(fnaive) + ggplot2::scale_y_log10()
}
\author{
Rob J Hyndman
}
