% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMOMO.R
\name{GAPC}
\alias{GAPC}
\alias{report.GAPC}
\title{Generalized APC stochastic mortality model}
\usage{
GAPC(formula, use_weights = TRUE, clip = 0, zeroCohorts = NULL, ...)
}
\arguments{
\item{formula}{Model specification}

\item{use_weights}{If \code{TRUE}, will call \code{\link[stMoMo]{genWeightMat}} with arguments \code{clip} and \code{zeroCohorts}.}

\item{clip}{Passed to \code{\link[stMoMo]{genWeightMat}()}}

\item{zeroCohorts}{Passed to \code{\link[stMoMo]{genWeightMat}()}}

\item{...}{All other arguments passed to \code{\link[StMoMo]{StMoMo}()}}
}
\value{
A model specification.
}
\description{
A Generalized Age-Period-Cohort (GAPC) stochastic mortality model, as defined
in Villegas et al. (2018). The StMoMo package is used to fit the model.
\code{GAPC()} returns a GAPC model applied to the formula's response variable as a function of age.
The model will optionally call \code{\link[stMoMo]{genWeightMat}} with arguments \code{clip} and \code{zeroCohorts}.
All other arguments are passed to \code{\link[StMoMo]{StMoMo}}.
}
\examples{
# CBD model
f2 <- function(x, ages) x - mean(ages)
cbd <- aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(cbd = GAPC(Mortality, link = "log",
    staticAgeFun = FALSE, periodAgeFun = c("1", f2))
  )
report(cbd)
}
\references{
Villegas, A.M., Millossovich, P., and Kaishev, V.K. (2018).
StMoMo: An R package for stochastic mortality modelling.
\emph{Journal of Statistical Software}, \strong{84}(3), 1-38.
\url{doi:10.18637/jss.v084.i03}
}
\author{
Rob J Hyndman
}
