% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/STMOMO.R
\name{GAPC}
\alias{GAPC}
\alias{report.GAPC}
\title{Generalized APC stochastic mortality model}
\usage{
GAPC(formula, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Arguments passed to StMoMo::StMoMo()}
}
\value{
A model specification.
}
\description{
A Generalized Age-Period-Cohort (GAPC) stochastic mortality model, as defined in the StMoMo package.
\code{GAPC()} returns a GAPC model applied to the formula's response
variable as a function of age.
}
\examples{
f2 <- function(x, ages) x - mean(ages)
cbd <- aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(cbd = GAPC(Mortality, link = "logit", staticAgeFun = FALSE,
  periodAgeFun = c("1", f2)))
report(cbd)
}
\references{
Villegas, A.M., Millossovich, P., and Kaishev, V.K. (2018).
StMoMo: An R package for stochastic mortality modelling.
\emph{Journal of Statistical Software}, \strong{84}(3), 1-38.
\url{doi:10.18637/jss.v084.i03}
}
\author{
Rob J Hyndman
}
