% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R, R/fmean.R, R/generate.R
\name{generate.LC}
\alias{generate.LC}
\alias{generate.FMEAN}
\alias{generate.mdl_vtl_df}
\title{Generate responses from a mable}
\usage{
\method{generate}{LC}(x, new_data, bootstrap = FALSE, ...)

\method{generate}{FMEAN}(x, new_data, h = NULL, bootstrap = FALSE, times = 1, seed = NULL, ...)

\method{generate}{mdl_vtl_df}(
  x,
  new_data = NULL,
  h = NULL,
  bootstrap = FALSE,
  times = 1,
  seed = NULL,
  bootstrap_block_size = 1,
  ...
)
}
\arguments{
\item{x}{A mable.}

\item{new_data}{The data to be generated (time index and exogenous regressors)}

\item{bootstrap}{If TRUE, then forecast distributions are computed using simulation with resampled errors.}

\item{...}{Additional arguments}

\item{h}{The simulation horizon (can be used instead of `new_data` for regular time series with no exogenous regressors).}

\item{times}{The number of replications.}

\item{seed}{The seed for the random generation from distributions.}

\item{bootstrap_block_size}{The bootstrap block size specifies the number of contiguous residuals to be taken in each bootstrap sample.}
}
\description{
Use a model's fitted distribution to simulate additional data with similar
behaviour to the response. This is a tidy implementation of
`\link[stats]{simulate}`.
}
\details{
Innovations are sampled by the model's assumed error distribution.
If `bootstrap` is `TRUE`, innovations will be sampled from the model's residuals.
}
\author{
Rob J Hyndman and Mitchell O'Hara-Wild
}
