% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popsim.R
\name{generate_population}
\alias{generate_population}
\title{Future population simulation}
\usage{
generate_population(
  starting_population,
  mortality_model,
  fertility_model,
  migration_model,
  h = 10,
  n_reps = 1000
)
}
\arguments{
\item{starting_population}{A \code{vital} object with the age-sex-specific starting population.}

\item{mortality_model}{A \code{mable} object containing an age-sex-specific model for mortality rates,
trained on data up to the year of the starting population.}

\item{fertility_model}{A \code{mable} object containing an age-specific model for fertility rates,
trained on data up to the year of the starting population.}

\item{migration_model}{A \code{mable} object containing an age-sex-specific model for net migration numbers,
trained on data up to the year of the starting population.}

\item{h}{The forecast horizon equal to the number of years to simulate into the future.}

\item{n_reps}{The number of replicates to simulate.}
}
\value{
A \code{vital} object containing the simulated future population.
}
\description{
Simulate future age-specific population given a starting population and
models for fertility, mortality, and migration. This is an experimental function
and has not been thoroughly tested.
}
