% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R, R/fmean.R
\name{interpolate.LC}
\alias{interpolate.LC}
\alias{interpolate.FMEAN}
\title{Interpolate missing values}
\usage{
\method{interpolate}{LC}(object, new_data, ...)

\method{interpolate}{FMEAN}(object, new_data, ...)
}
\arguments{
\item{object}{A mable containing a single model column.}

\item{new_data}{A dataset with the same structure as the data used to fit the model.}

\item{...}{Other arguments passed to interpolate methods.}
}
\description{
Uses a fitted model to interpolate missing values from a dataset.

Uses a fitted model to interpolate missing values from a dataset.
}
\author{
Rob J Hyndman
}
