% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R, R/fmean.R
\name{interpolate.model_lc}
\alias{interpolate.model_lc}
\alias{interpolate.model_fmean}
\title{Interpolate missing values}
\usage{
\method{interpolate}{model_lc}(object, new_data, ...)

\method{interpolate}{model_fmean}(object, new_data, ...)
}
\arguments{
\item{object}{A mable containing a single model column.}

\item{new_data}{A dataset with the same structure as the data used to fit the model.}

\item{...}{Other arguments passed to interpolate methods.}
}
\description{
Uses a fitted model to interpolate missing values from a dataset.

Uses a fitted model to interpolate missing values from a dataset.
}
