% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average.R
\name{interpolate.model_ave}
\alias{interpolate.model_ave}
\title{Interpolate missing values}
\usage{
\method{interpolate}{model_ave}(object, new_data, ...)
}
\arguments{
\item{object}{A mable containing a single model column.}

\item{new_data}{A dataset with the same structure as the data used to fit the model.}

\item{...}{Other arguments passed to interpolate methods.}
}
\description{
Uses a fitted model to interpolate missing values from a dataset.
}
