% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R
\name{LC}
\alias{LC}
\alias{report.model_lc}
\title{Lee-Carter models}
\usage{
LC(formula, adjust = c("dt", "dxt", "e0", "none"), scale = FALSE, ...)
}
\arguments{
\item{formula}{Model specification.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
\code{LC()} returns a Lee-Carter model applied to the formula's response variable as a function of age.
}
\examples{
library(dplyr)
aus_mortality |>
  filter(State == "Victoria", Sex == "female") |>
  model(lee_carter = LC(Mortality))
}
