% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC.R
\name{LC}
\alias{LC}
\alias{report.LC}
\title{Lee-Carter model}
\usage{
LC(
  formula,
  adjust = c("dt", "dxt", "e0", "none"),
  jumpchoice = c("fit", "actual"),
  scale = FALSE,
  ...
)
}
\arguments{
\item{formula}{Model specification.}

\item{adjust}{method to use for adjustment of coefficients \eqn{k_t kt}.
Possibilities are
\dQuote{dt} (Lee-Carter method, the default),
\dQuote{dxt} (BMS method),
\dQuote{e0} (Lee-Miller method based on life expectancy) and
\dQuote{none}.
y).}

\item{jumpchoice}{Method used for computation of jumpchoice. Possibilities: “actual” (use actual rates from final year) and “fit” (use fitted rates).}

\item{scale}{If TRUE, bx and kt are rescaled so that kt has drift parameter = 1.}

\item{...}{Not used.}
}
\value{
A model specification.
}
\description{
Lee-Carter model of mortality or fertility rates.
\code{LC()} returns a Lee-Carter model applied to the formula's response
variable as a function of age. This produces a standard Lee-Carter model by
default, although many other options are available. Missing rates are set to
the geometric mean rate for the relevant age.
}
\examples{
lc <- aus_mortality |>
  dplyr::filter(State == "Victoria", Sex == "female") |>
  model(lee_carter = LC(Mortality))
report(lc)
autoplot(lc)
}
\references{
Basellini, U, Camarda, C G, and Booth, H (2022) Thirty years on:
A review of the Lee-Carter method for forecasting mortality.
\emph{International Journal of Forecasting}, \bold{39}(3), 1033-1049.
\url{https://doi.org/10.1016/j.ijforecast.2022.11.002}

Booth, H., Maindonald, J., and Smith, L. (2002) Applying Lee-Carter
under conditions of variable mortality decline. \emph{Population Studies},
\bold{56}, 325-336. \url{https://doi.org/10.1080/00324720215935}

Lee, R D, and Carter, L R (1992) Modeling and forecasting US mortality.
\emph{Journal of the American Statistical Association}, \bold{87}, 659-671.
\url{https://doi.org/10.1080/01621459.1992.10475265}

Lee R D, and Miller T (2001). Evaluating the performance of the Lee-Carter
  method for forecasting mortality. \emph{Demography}, \bold{38}(4), 537–549.
\url{https://doi.org/10.1353/dem.2001.0036}
}
\author{
Rob J Hyndman
}
