% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lee_carter.R
\name{lee_carter}
\alias{lee_carter}
\title{Model mortality or fertility data using Lee-Carter approach}
\usage{
lee_carter(
  .data,
  age,
  sex,
  rates,
  pop,
  adjust = c("dt", "dxt", "e0", "none"),
  scale = FALSE
)
}
\arguments{
\item{.data}{A tsibble including an age variable and a variable containing mortality or fertility rates.}

\item{age}{Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` or `Age_group` will be used (not case sensitive).}

\item{sex}{Optional variable in `.data` containing sex information. If omitted, the variable with name `Sex` or `Group` will be used (not case sensitive).}

\item{rates}{Variable in `.data` containing mortality or fertility rates.
If omitted, it will search for a variable with one of the following names:
`mx`, `mortality`, `fx`, `fertility` or `rate` (not case sensitive).}

\item{pop}{Variable in `.data` containing population numbers. If omitted, it
will search for a variable with one of the following names:
`pop`, `population`, `ex`, `exposure`.}

\item{adjust}{method to use for adjustment of coefficients \eqn{k_t kt}.
Possibilities are
\dQuote{dt} (Lee-Carter method, the default),
\dQuote{dxt} (BMS method),
\dQuote{e0} (Lee-Miller method based on life expectancy) and
\dQuote{none}.}

\item{scale}{If TRUE, bx and kt are rescaled so that kt has drift parameter = 1.}
}
\value{
A list of class \code{lc_model} containing various model objects.
}
\description{
Lee-Carter model of mortality or fertility rates. \code{lee_carter} produces a
standard Lee-Carter model by default, although many other options are
available. Missing rates are set to the geometric mean rate for the relevant age.
}
\examples{
# Compute Lee-Carter model for Australian females, males and total
aus_lc <- aus_mortality |>
  dplyr::filter(Code == "AUS") |>
  lee_carter()
aus_lc
autoplot(aus_lc, "Lee Carter components for Australia")
autoplot(aus_lc$time, kt)
}
\references{
Basellini, U, Camarda, C G, and Booth, H (2022) Thirty years on:
A review of the Lee-Carter method for forecasting mortality.

Booth, H., Maindonald, J., and Smith, L. (2002) Applying Lee-Carter
under conditions of variable mortality decline. \emph{Population Studies}, \bold{56}, 325-336.

Lee, R D, and Carter, L R (1992) Modeling and forecasting US mortality.
\emph{Journal of the American Statistical Association}, \bold{87}, 659-671.

Lee R D, and Miller T (2001). Evaluating the performance of the Lee-Carter
  method for forecasting mortality. \emph{Demography}, \bold{38}(4), 537–549.
\emph{International Journal of Forecasting}, to appear.
}
\seealso{
\code{\link{forecast.lc_model}()}
}
\author{
Rob J Hyndman.
}
\keyword{models}
