% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table.R
\name{life_table}
\alias{life_table}
\title{Compute period lifetables from mortality rates}
\usage{
life_table(.data, age, mortality, group)
}
\arguments{
\item{data.}{A tsibble containing mortality rates, or deaths and exposures. It must contain an age variable and a variable containing mortality rates.}

\item{age.}{Variable in `data` containing start year of age intervals.}

\item{mortality.}{Variable in `data` containing Mortality rates (mx).}
}
\value{
A tsibble object.
}
\description{
All available years and ages are included in the tables.
qx = mx/(1 + ((1-ax) * mx)) as per Chiang (1984).
Warning: the code has only been tested for data based on single-year age groups.
}
\examples{
# Compute French life tables
demography::fr.mort \%>\%
  as_tsibble() \%>\%
  life_table(age=Age, mortality=Mortality, group=Group)
}
\references{
Chiang CL. (1984) \emph{The life table and its applications}. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) \emph{Applied mathematical demography}, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) \emph{Demography: measuring and modeling population processes}. Blackwell
}
