% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/life_table.R
\name{life_table}
\alias{life_table}
\title{Compute period lifetables from mortality rates}
\usage{
life_table(.data, age, sex, mortality)
}
\arguments{
\item{.data}{A tsibble containing mortality rates, or deaths and exposures. It must contain an age variable and a variable containing mortality rates.}

\item{age}{Variable in `.data` containing start year of age intervals. If omitted, the variable with name `Age` will be used (not case sensitive).}

\item{sex}{Variable in `.data` containing sex information. If omitted, the variable with name `Sex` or `Group` will be used (not case sensitive).}

\item{mortality}{Variable in `.data` containing Mortality rates (mx). If omitted, the variable with name  `Mortality` or `Rate` will be used (not case sensitive).}
}
\value{
A tsibble object.
}
\description{
All available years and ages are included in the tables.
qx = mx/(1 + ((1-ax) * mx)) as per Chiang (1984).
Warning: the code has only been tested for data based on single-year age groups.
}
\examples{
# Compute Australia life table for females in 2003
aus_mortality \%>\%
  filter(Code=="AUS", Sex=="female", Year==2003) \%>\%
  life_table()
}
\references{
Chiang CL. (1984) \emph{The life table and its applications}. Robert E Krieger Publishing Company: Malabar.

Keyfitz, N, and Caswell, H. (2005) \emph{Applied mathematical demography}, Springer-Verlag: New York.

Preston, S.H., Heuveline, P., and Guillot, M. (2001) \emph{Demography: measuring and modeling population processes}. Blackwell
}
