% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pr.R
\name{make_pr}
\alias{make_pr}
\title{Make a new vital containing products and ratios from a key variable}
\usage{
make_pr(.data, key = "Sex")
}
\arguments{
\item{.data}{A vital object}

\item{key}{A character string specifying the key variable}
}
\value{
A vital object
}
\description{
Make a new vital containing products and ratios from a key variable
}
