% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_migration.R
\name{net_migration}
\alias{net_migration}
\title{Calculate net migration from a vital object}
\usage{
net_migration(deaths, births)
}
\arguments{
\item{deaths}{A vital object containing at least the columns Year, Population and Deaths}

\item{births}{A vital object containing at least the columns Year and Births}
}
\value{
A vital object containing the same columns as deaths plus Net Migration computed
using the formula Net Migration = Population - lag(Population) - Deaths + Births
}
\description{
Calculate net migration from a vital object
}
