% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{read_hmd}
\alias{read_hmd}
\title{Read data from the Human Mortality Database}
\usage{
read_hmd(country)
}
\arguments{
\item{country}{Character string containing the country code. It may be
a named or unnamed character vector. If the vector is named, the name is used
as the label.}
}
\value{
A tsibble with an annual index, and Age, Sex and Country keys.
}
\description{
Download and merge data from the Human Mortality Database (\url{http://mortality.org}).
It returns births, deaths, population, exposure, and death rates for single
years of age (i.e., the 1x1 files). For other types of data extraction,
use the `HMDHFDplus` package.
}
\details{
Before running this function,
you need to register at the "former" HMD website (https://former.mortality.org),
and use \code{\link{hmd_session}()} to register your username and password.
Usernames and passwords for the new HMD site (https://mortality.org) will not
work. This is a temporary patch until an API is released for the new site.
}
\examples{
\dontrun{
aus <- read_hmd("AUS")
}
}
\seealso{
\code{\link[HMDHFDplus]{readHMD}}.
}
\author{
Emi Tanaka and Rob Hyndman
}
