% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmd.R
\name{read_hmd}
\alias{read_hmd}
\title{Read data directly from HMD and construct a \code{vital} object for use in other functions}
\usage{
read_hmd(
  country,
  username,
  password,
  variables = c("Deaths", "Exposures", "Population", "Mx")
)
}
\arguments{
\item{country}{Directory abbreviation from the HMD. For instance, Australia = "AUS".}

\item{username}{HMD username (case-sensitive)}

\item{password}{HMD password (case-sensitive)}

\item{variables}{List of variables to download from the HMD.}
}
\value{
\code{read_hmd} returns a \code{vital} object with the following variables:
\item{Year}{Numerical vector containing year of observation}
\item{Age}{Numerical vector containing age group}
\item{OpenInterval}{Logical vector indicating if the age interval is open}
\item{Sex}{Character vector taking values "Female", "Male", "Total"}
\item{Deaths}{Total deaths}
\item{Exposure}{Numerical vector of exposure-to-risk, equal to population size on 30 June}
\item{Population}{Numerical vector containing population size on 1 January}
\item{Mortality}{Numerical vector of mortality rate, equal to Deaths/Exposure}
}
\description{
\code{read_hmd} reads 1x1 data from the Human Mortality Database (HMD
\url{https://www.mortality.org}) and constructs a \code{vital} object suitable
for use in other functions. This function uses \code{\link[HMDHFDplus:readHMDweb]{HMDHFDplus::readHMDweb()}}
to download the required data.
}
\details{
In order to read the data, users are required to create an account with the
HMD website (\url{https://www.mortality.org}), and obtain a valid username and password.
}
\examples{
\dontrun{
norway <- read_hmd(
  country = "NOR",
  username = "Nora.Nilsen@mymail.com",
  password = "FF!5xeEFa6"
)
}
}
\author{
Rob J Hyndman
}
