% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ktdb.R
\name{read_ktdb_file}
\alias{read_ktdb_file}
\title{#' Read old-age mortality data from files downloaded from K-T database}
\usage{
read_ktdb_file(file, triangle = 1)
}
\arguments{
\item{file}{Name of a file containing data downloaded from the K-T database.}

\item{triangle}{Lexis triangle number, 1 (default) is lower triangle, 2 is upper triangle.}
}
\value{
\code{read_ktdb_file} returns a \code{vital} object combining the downloaded data.
}
\description{
\code{read_ktdb_file} reads old-age mortality data from a file downloaded from
K-T database (\url{https://www.demogr.mpg.de/cgi-bin/databases/ktdb/datamap.plx})
and constructs a \code{vital} object suitable for use in other functions.
}
\examples{
\dontrun{
# File downloaded from the K-T database
australia_male <- read_ktdb_file("maustl.txt")
}

}
\author{
Sixian Tang
}
