% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{forecast}
\alias{generate}
\alias{model}
\alias{report}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabletools}{\code{\link[fabletools]{model}}, \code{\link[fabletools]{report}}}

  \item{generics}{\code{\link[generics]{forecast}}, \code{\link[generics]{generate}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

