% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{autoplot}
\alias{forecast}
\alias{generate}
\alias{model}
\alias{report}
\alias{interpolate}
\alias{tidy}
\alias{augment}
\alias{estimate}
\alias{glance}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{fabletools}{\code{\link[fabletools:reexports]{augment}}, \code{\link[fabletools]{estimate}}, \code{\link[fabletools:reexports]{glance}}, \code{\link[fabletools:reexports]{interpolate}}, \code{\link[fabletools]{model}}, \code{\link[fabletools]{report}}, \code{\link[fabletools:reexports]{tidy}}}

  \item{generics}{\code{\link[generics]{forecast}}, \code{\link[generics]{generate}}}

  \item{ggplot2}{\code{\link[ggplot2]{autoplot}}}
}}

