% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth_vital}
\alias{smooth_vital}
\alias{smooth_spline}
\alias{smooth_mortality}
\alias{smooth_fertility}
\alias{smooth_loess}
\title{Functions to smooth demographic data}
\usage{
smooth_vital(data, ...)

smooth_spline(x, y, k = -1, weights = NULL)

smooth_mortality(x, y, b = 65, power = 0.4, weights = NULL)

smooth_fertility(x, y, b = 65, k = 30, lambda = 1e-10)

smooth_loess(x, y, span = 0.2, weights = NULL)
}
\arguments{
\item{x}{vector of ages}

\item{y}{vector to smooth (e.g., mortality or fertility rates)}

\item{k}{Number of knots to use for penalized regression spline estimate.}

\item{weights}{Vector of weights.}

\item{b}{Lower age for monotonicity. Above this, the smooth curve is assumed to be monotonically increasing.}

\item{power}{Power transformation for age variable before smoothing. Default is 0.4 for mortality data and 1 (no transformation) for fertility or migration data.}

\item{lambda}{Penalty for constrained regression spline.}

\item{span}{Span for loess smooth.}
}
\value{
Vector of smoothed values
}
\description{
`smooth_vital()` allows smoothing of a vital object. It effectively groups
the data by keys other than age, and the index, and then applies the smoothing
functions specified. So it is like a group_by() followed by a mutate(), where the
grouping is determined by the attributes of the vital object.
}
\details{
`smooth_spline()` and `smooth_mortality()` use penalized regression splines, the latter with a
monotonicity constraint above age `b`. The methodology is based on Wood (1994).
`smooth_fertility()` uses weighted regression B-splines with a concavity constraint,
based on He and Ng (1999). Finally, `smooth_loess()` uses locally quadratic
regression.
}
\examples{
library(dplyr)
aus_mortality |>
  smooth_vital(sm_mort = smooth_mortality(Age, Mortality))
}
\author{
Rob J Hyndman
}
\keyword{smooth}
