% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmean.R
\name{tidy.model_fmean}
\alias{tidy.model_fmean}
\title{Extract model coefficients from a mable}
\usage{
\method{tidy}{model_fmean}(x, ...)
}
\arguments{
\item{x}{A mable.}

\item{...}{Arguments for model methods.}
}
\description{
This function will obtain the coefficients (and associated statistics) for
each model in the mable.
}
