% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_package.R
\docType{package}
\name{vital-package}
\alias{vital-package}
\title{vital: Tidy Analysis Tools for Mortality, Fertility, Migration and Population Data}
\description{
Functions for analysing vital statistics based on tools consistent with the tidyverse. Tools are provided for data visualization, lifetable calculations, computing net migration numbers, Lee-Carter modelling; functional data modelling and stochastic population forecasting.
}
\examples{
# create a vital with only age as a key
vital(
  year = rep(2010:2015, 100),
  age = rep(0:99, each = 6),
  mx = runif(600, 0, 1),
  index = year,
  key = age,
  .age = age
)
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/vital}
  \item \url{http://github.com/robjhyndman/vital}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

}
\keyword{internal}
