% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_package.R
\docType{package}
\name{vital-package}
\alias{vital-package}
\title{vital: Tidy Analysis Tools for Mortality, Fertility, Migration and Population Data}
\description{
Analysing vital statistics based on tools consistent with the tidyverse. Tools are provided for data visualization, life table calculations, computing net migration numbers, Lee-Carter modelling; functional data modelling and forecasting.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/vital/}
  \item \url{https://github.com/robjhyndman/vital}
  \item Report bugs at \url{https://github.com/robjhyndman/vital/issues}
}

}
\author{
\strong{Maintainer}: Rob Hyndman \email{Rob.Hyndman@monash.edu} (\href{https://orcid.org/0000-0002-2140-5352}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Sixian Tang (\href{https://orcid.org/0000-0002-7292-462X}{ORCID})
}

Other contributors:
\itemize{
  \item Miles McBain (\href{https://orcid.org/0000-0003-2865-2548}{ORCID}) [contributor]
  \item Mitchell O'Hara-Wild (\href{https://orcid.org/0000-0001-6729-7695}{ORCID}) [contributor]
}

}
\keyword{internal}
