% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vital_vars.R
\name{vital_vars}
\alias{vital_vars}
\title{Return vital variables}
\usage{
vital_vars(x)
}
\arguments{
\item{x}{A tsibble object.}
}
\value{
A character vector of the names of the vital variables.
}
\description{
A vital object is a special case of a tsibble object with additional attributes
identifying the age, sex, deaths, births and population variables.
\code{vital_vars()} returns a character vector the names of the vital variables.
}
\examples{
vital_vars(norway_mortality)

}
